Update controls.md
The Hide NoControls warning snippet is incorrect so the description was a bit challenging to understand as a new user. I don't think it would hurt to make it more explicit. I will attempt to update the incorrect snippet as well.

The outline hierarchy is also off as *Hide NoControls* and *Show full documentation* follow from *Configurations#Parameter*. Not sure if this change conflicts with any style guidance for level 4 headings.